LinuxTarget: Allow determining target ABI without busybox
We need to know the target ABI in order to install busybox, therefore we
should attempt to find the ABI when busybox is not yet present. I'm not
aware of a clean way to do this if there's no `uname` command present,
so just raise an exception in that case.