gem5: Set the kernel log level to 1
When running with gem5, we use the same virtual terminal as is used
for the kernel log. This can cause issues when the kernel prints, as
we might be trying to parse the output from a command. This can cause
runs to crash. We therefore set the kernel log level to 1 by default,
which should supress all log messages aside from ones such as kernel
panics. This setting is changed as the first thing once we have
established a reliable connection with the simulated device.