instrument/schedstats: Add instrument for schedstats
This instrument collects snapshots of the CPU/sched_domain scheduler
statistics in /proc/schedstat, and reports a Measure for each of
them. It currently doesn't support the per-task stats reported in
/proc/<pid>/schedstat, and it ignores the additional EAS schedstats
reported by Android kernels.

The measures are named according to the corresponding identifiers in
the kernel source code. Although This might appear to be an obscure
way to refer to them, the information in /proc/schedstats is not
going to be any use to anyone who doesn't have the kernel code
handy. There is an English description in
Documentation/scheduler/sched-stats.txt but it can only really be
considered a reference for those who already understand the
scheduler.

Some of the measures are representing time intervals, but their units
are jiffies, the meaning of which depends on the kernel
configuration. Therefore no units are reported for the measurements.