cgroup: wait to verify value in set()
The standard behavior for write_value() is to read it back immediately
afterwards and verify that it has been written. When writing some cgroup
parameter values (e.g. "FROZEN"), it may take a while for them to take
effect. If they are read back immediately afterwards, the result won't
match the expected value, producing an error.

While it is possible to disable validation on write, it is still
desirable to make sure that the parameters are set properly. This commit
changes the behavior to continue reading the value up to a timeout
period afterwards, giving it time to take effect, and only raise an
error if the timeout is reached.