instrument/frames: Fix parsing MeasurementsCsv
The collector's write_frames method uses a fixed column order, only using the
'columns' param as a filter i.e. if its columns are ['a', 'b', 'c'] and you pass
columns=['c', 'a'], the output columns will be ['a', 'c'].

Therefore passing self.active_channels breaks MeasurementsCsv parsing when its
order does not match the collector's column order.

Since the collector writes a header in the CSV file, just pass columns=None to
MeasurementsCsv so it will parse that header to detect the column order.