android: Add AndroidScreen class with orientation methods
This can be used to save, set and restore the screen
orientation.

This can be useful for ensuring the reproducibility of results that
depend on the actual content that was rendered. A simple example of
this is that a device with an OLED screen uses measurably more energy
playing a plain-white Youtube video if it is in landscape orentation,
due to having more white pixels lit up.