utils/android: Append platform-tools at the end of PATH
devlib currently prepends platform-tools at the beginning of PATH.
This makes WA3 use a different adb (the one in platform-tools) than
the one I use everywhere else.  Because they are different, whenever
the "other" adb starts, it kills the sessions of the previous one.

Instead of prepending and using a different adb than the reset of the
system, append platform-tools to PATH, so that the adb used by WA3 and
the rest of the system is the same.