Allow connections to be used as context managers
Add a base class for connection objects that implements the context
manager interface and keeps track of the connection's target (which gets
set by Target.get_connection). When used as a context manager, the connection
substitutes itself for the target's default connection for the thread
within the context using Target.set_connect(). The old connection is
restored upon exiting the context.