android: Fix adb background command PID detection
Sometimes the reported PID for a background command would be reported
incorrectly. This was due to an assumption that the lowest PID
was always the command we were looking for, however this is not always
true and PID's could be recycled resulting in the PID of the search
command being reported instead.

To avoid this, echo the PID directly using $$ similar to how the
SSH connection handles this scenario.