target: Add Target.revertable_write_value(may_not_exist=False)
Add a may_not_exist parameter to Target.revertable_write_value() that
will turn the write into a no-op if the file does not exist or cannot be
read. If may_not_exist is a callable, it will be called with the
parameters of the function. This can be used to e.g. log the issue at a
higher logging level.